GH-1188: Improved tests
