+++ projects/storefrontlib/src/lib/my-account/close-account/components/close-account/close-account.component.spec.ts
@@ -1,10 +1,9 @@
 import { async, TestBed, ComponentFixture } from '@angular/core/testing';
 import { CloseAccountComponent } from './close-account.component';
-import { I18nTestingModule, UserToken, AuthService } from '@spartacus/core';
+import { I18nTestingModule } from '@spartacus/core';
 import { RouterTestingModule } from '@angular/router/testing';
 import { Pipe, PipeTransform } from '@angular/core';
 import { NgbModule, NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { Observable, of } from 'rxjs';
 import { CloseAccountModalComponent } from '../close-account-modal/close-account-modal.component';
 
 @Pipe({
@@ -14,29 +13,16 @@ class MockTranslateUrlPipe implements PipeTransform {
   transform(): any {}
 }
 
-class MockAuthService {
-  getUserToken(): Observable<UserToken> {
-    return of();
-  }
-}
-
 describe('CloseAccountComponent', () => {
   let component: CloseAccountComponent;
   let fixture: ComponentFixture<CloseAccountComponent>;
   let modalInstance: any;
-  let authService: AuthService;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [I18nTestingModule, NgbModule, RouterTestingModule],
       declarations: [CloseAccountComponent, MockTranslateUrlPipe],
-      providers: [
-        {
-          provide: AuthService,
-          useClass: MockAuthService,
-        },
-        { provide: NgbModal, useValue: { open: () => {} } },
-      ],
+      providers: [{ provide: NgbModal, useValue: { open: () => {} } }],
     }).compileComponents();
   }));
 
@@ -44,7 +30,6 @@ describe('CloseAccountComponent', () => {
     fixture = TestBed.createComponent(CloseAccountComponent);
     component = fixture.componentInstance;
     modalInstance = TestBed.get(NgbModal);
-    authService = TestBed.get(AuthService);
 
     spyOn(modalInstance, 'open').and.returnValue({ componentInstance: {} });
     fixture.detectChanges();
@@ -54,16 +39,12 @@ describe('CloseAccountComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should open modal an pass userToken', () => {
-    spyOn(authService, 'getUserToken').and.returnValue({
-      userId: 'userId1',
-    } as UserToken);
+  it('should open modal', () => {
     component.openModal();
 
     expect(modalInstance.open).toHaveBeenCalledWith(
       CloseAccountModalComponent,
       Object({ centered: true })
     );
-    expect(component.modal.userToken$).toEqual({ userId: 'userId1' });
   });
 });
